Cookies setting and getting & demo
The only real change in EWC is to add SetCookie and GetCookie.

The demo shows setting, refreshing (in case cookies have been
changed in the dev console, and clearing a cookie. Best way to
test is to set a cookie and close things down and then come back
again and see that the demo remembers.